requirements.txt
# Note: no dependencies are required on Python >=3.9; `black` is optional; `pyyaml` is mostly optional
https://github.com/offscale/astor/archive/refs/heads/empty-annassign.zip
python-cdd>=0.0.84b0
black
pyyaml
typing-extensions
pyarrow
